updated dag
