+++ dags/openshift_nightlies/dag.py
@@ -105,16 +105,32 @@ class BaremetalOpenshiftNightlyDAG(AbstractOpenshiftNightlyDAG):
         webfuse_installer = self._get_webfuse_installer()
 
         install_cluster = bm_installer.get_install_task()
-        deploy_webfuse = webfuse_installer.get_deploy_app_task()
-        scaleup_cluster = bm_installer.get_scaleup_task()
+        benchmark_stg_1 = self._add_benchmarks(task_group="install-bench")
 
-        install_cluster >> scaleup_cluster >> deploy_webfuse
+        deploy_webfuse = webfuse_installer.get_deploy_app_task()
+        benchmark_stg_2 = self._add_benchmarks(task_group="scaleup-bench")
+
+        scaleup_cluster = bm_installer.get_scaleup_task()
+        benchmark_stg_3 = self._add_benchmarks(task_group="webfuse-bench")
+
+        install_cluster >> benchmark_stg_1 
+        install_cluster >> scaleup_cluster >> benchmark_stg_2 
+        scaleup_cluster >> deploy_webfuse >> benchmark_stg_3
 
     def _get_openshift_installer(self):
         return jetski.BaremetalOpenshiftInstaller(self.dag, self.config, self.release)
 
     def _get_webfuse_installer(self):
         return webfuse.BaremetalWebfuseInstaller(self.dag, self.config, self.release)
+
+    def _get_e2e_benchmarks(self, task_group):
+        return e2e.E2EBenchmarks(self.dag, self.config, self.release, task_group)
+
+    def _add_benchmarks(self, task_group):
+        with TaskGroup(task_group, prefix_group_id=True, dag=self.dag) as benchmarks:
+            benchmark_tasks = self._get_e2e_benchmarks(task_group).get_benchmarks()
+            chain(*benchmark_tasks)
+        return benchmarks
 
 class OpenstackNightlyDAG(AbstractOpenshiftNightlyDAG):
     def build(self):
